Requests merge step 1: rename the tracker, recover the id space
The Shared Deeds tracker moves to /requests. /shared-deeds stays as a
permanent alias — ?focus= links to it are in mail already delivered, and
an email in somebody's inbox cannot be edited.

THE ID SPACE. Two pages became one and their rows come from different
tables: a review is a deed_shares.id, a signing is a
signing_requests.id. Review 42 and signing 42 both exist and are
different deeds belonging to different people, so ?focus=42 on the
merged page names two rows and the path that used to disambiguate it is
gone. The alias supplies what the number cannot — an id arriving at
/shared-deeds came from the reviews table, because that is the only
thing that path has ever meant, and it leaves saying so.

A ?focus= with no ?kind= is genuinely ambiguous and refuses: filter
stays on "all", nothing is highlighted. Guessing "reviews first" would
be right most of the time and silently highlight a stranger's signing
the rest — a tie-breaking rule invents an answer that will be right
often enough that nobody checks it.

The rule lives in lib/requestsFocus.ts rather than in expressions in the
page, because what matters here is NEGATIVE — an ambiguous link
highlights nothing — and source text cannot demonstrate the absence of a
guess. A test has to call it and read the answer.

The four backend links that built /shared-deeds?focus= now build
/requests?kind=reviews&focus=. The alias is permanent because sent mail
cannot be edited, and that covers exactly the mail already sent;
continuing to emit the old path would grow the legacy population forever
and make the alias self-justifying. Two pins hold it.

Caught on the way past: apiFetch('/shared-deeds', {method:'POST'}) in
the review modal was repointed with the navigation and would have 404'd
every review share on the first click. The page and the endpoint share a
name; that trap now has a pin.

The auth-guard sweep flagged the alias, correctly. Rather than a fourth
hand-listed exemption, redirect stubs are now recognised by a classifier
that PROVES the claim — imports redirect, calls it, and contains no
fetch, token read, storage, state or markup — with floor assertions at
both ends and a mutation probe in the suite.

Pins retargeted with their reasoning at the read site: six frontend
suites and the cross-language contract pin in Python, which announced
the move by raising on a file with no interface in it rather than
passing quietly.

Deferred to step 2: folding in the agenda's grouping, moving the
expandable signing detail across, retiring /signings to an alias. The
two row shapes stay different, and must.

Gates: pytest 1156, jest 779/53 suites, tsc 88, six-flow green,
banned-claims clean, next build clean. Eight mutation probes, all caught.
